@@ -1,6 +1,9 @@
 source 'https://rubygems.org'
+ruby '2.3.0'
 
 gem 'active_record_union'
+gem 'amatch'
+gem 'aws-sdk'
 gem 'bourbon'
 gem 'canard', '0.4.2.pre'
 gem 'capistrano-bundler', require: false
@@ -9,48 +12,45 @@ gem 'capistrano-rails',   require: false
 gem 'capistrano-rvm',     require: false
 gem 'capistrano',         require: false
 gem 'capistrano3-puma',   require: false
-gem 'underscore-rails'
 gem 'codeclimate-test-reporter', group: :test, require: nil
-gem 'coffee-rails', '~> 4.0.0'
+gem 'coffee-rails'
 gem 'coveralls', require: false
-gem 'dropzonejs-rails'
+gem 'dogapi', '>=1.3.0'
 gem 'dotenv-rails'
-gem 'haml-rails'
-gem 'inline_svg'
-gem 'premailer-rails'
+gem 'dropzonejs-rails'
 gem 'font-awesome-sass', '~> 4.5.0'
+gem 'fractional'
+gem 'geokit-rails'
+gem 'haml-rails'
 gem 'happy_path'
+gem 'ingreedy', git: 'https://github.com/danReynolds/ingreedy.git'
+gem 'inline_svg'
 gem 'jbuilder', '~> 2.0'
-gem 'jquery-datatables-rails', '~> 3.1.1'
 gem 'jquery-rails'
+gem 'linguistics'
 gem 'maildown'
 gem 'materialize-sass', git: 'https://github.com/danReynolds/materialize-sass.git'
+gem 'momentjs-rails'
 gem 'money-rails'
 gem 'mysql2'
 gem 'neat'
+gem 'newrelic_rpm'
 gem 'nokogiri'
+gem 'paperclip'
 gem 'puma'
-gem 'rails', '4.2.3'
-gem 'react-rails', '~> 1.5.0'
+gem 'rails', '~> 5.0.0'
+gem 'react-rails'
+gem 'rollbar'
+gem 'ruby-units'
 gem 'sass-rails'
 gem 'sdoc', '~> 0.4.0',          group: :doc
-gem 'simple_form'
-gem 'sorcery'
-gem 'uglifier', '>= 1.3.0'
-gem 'paperclip'
-gem 'aws-sdk', '< 2.0'
-gem 'linguistics'
-gem 'geokit-rails'
-gem 'rollbar'
-gem 'dogapi', '>=1.3.0'
-gem 'newrelic_rpm'
 gem 'sentry-raven'
-gem 'amatch'
-gem 'ruby-units'
-gem 'fractional'
-gem 'ingreedy', git: 'https://github.com/danReynolds/ingreedy.git'
-gem 'momentjs-rails'
+gem 'simple_form'
 gem 'slack-ruby-client'
+gem 'sorcery'
+gem 'turbolinks', '~> 5.0.0'
+gem 'uglifier', '>= 1.3.0'
+gem 'underscore-rails'
 
 if ENV["LOCAL_ENVIRONMENT"] == "development"
 	TESSERACT_GEM = 'git@github.com:danReynolds/ruby-tesseract-ocr.git'
@@ -67,7 +67,6 @@ end
 group :development, :test do
 	gem 'better_errors'
 	gem 'binding_of_caller'
-	gem 'meta_request'
 	gem 'stackprof'
   gem 'pry'
   gem 'pry-stack_explorer'
@@ -77,8 +76,8 @@ group :development, :test do
   gem 'rspec-rails'
   gem 'spring-commands-rspec'
   gem 'spring'
-	gem 'quiet_assets'
 	gem 'bullet'
 	gem 'gemsurance'
 	gem 'rack-mini-profiler'
+	gem 'rubocop'
 end
