@@ -1,5 +1,5 @@
 class ApplicationController < ActionController::Base
-  before_filter :require_login
+  before_action :require_login
   protect_from_forgery with: :exception
   extend HappyPath
   setup_happy_path
