@@ -1,4 +1,4 @@
-class GroceryPayment < ActiveRecord::Base
+class GroceryPayment < ApplicationRecord
   belongs_to :user
   belongs_to :grocery
   validates_uniqueness_of :grocery_id, scope: :user_id
