@@ -1,4 +1,4 @@
-class Recipe < ActiveRecord::Base
+class Recipe < ApplicationRecord
   has_many :items_recipes, class_name: ItemsRecipes
   has_and_belongs_to_many :items, through: :items_recipes
   has_and_belongs_to_many :groceries
