@@ -1,12 +1,12 @@
-class UserGroup < ActiveRecord::Base
+class UserGroup < ApplicationRecord
   has_many :user_groups_users, class_name: UserGroupsUsers
   has_many :users, through: :user_groups_users
-  has_many :items, -> { uniq }, through: :groceries
+  has_many :items, -> { distinct }, through: :groceries
   has_many :groceries
   has_many :payments, class_name: UserPayment
   has_many :user_defaults, class_name: User, foreign_key: :user_group_default_id
   has_many :slack_messages, through: :slack_bot
-  has_attached_file :banner, styles: { large: '800x600>', standard: '600x450>' }, default_url: 'user_groups/default2.jpg'
+  has_attached_file :banner, styles: { large: '800x600>', standard: '600x450>' }, default_url: 'user_groups/default1.jpg'
   has_one :slack_bot
   belongs_to :owner, class_name: User
 
