@@ -1,4 +1,4 @@
-class UserGroupsUsers < ActiveRecord::Base
+class UserGroupsUsers < ApplicationRecord
   validates_uniqueness_of :user_group_id, scope: :user_id
   belongs_to :user
   belongs_to :user_group
