@@ -1,2 +1,2 @@
 # config/initializers/paperclip.rb
-Paperclip::Attachment.default_options[:s3_host_name] = 's3-us-west-1.amazonaws.com'
+# Paperclip::Attachment.default_options[:s3_host_name] = 's3-us-west-1.amazonaws.com'
